Adding basic Tab completion.

+++ Debug/print/print.h
@@ -109,7 +109,7 @@ void int16ToStr ( uint16_t in, char*  out );
 void hexToStr_op( uint16_t in, char*  out, uint8_t op );
 void revsStr    ( char*  in );
 uint16_t lenStr ( char*  in );
-uint8_t eqStr   ( char*  str1, char* str2 ); // Returns 1 if identical, 0 otherwise
+int16_t eqStr   ( char*  str1, char* str2 ); // Returns -1 if identical, last character of str1 comparison (0 if str1 is like str2)
 
 #endif
